Updated snapshot tests to validate snapshots in old formats
