+++ packages/drivers/replay-driver/src/replayController.ts
@@ -23,7 +23,7 @@ export abstract class ReadDocumentStorageServiceBase implements IDocumentStorage
         return Promise.reject(new Error("Invalid operation"));
     }
 
-    public async write(tree: api.ITree, parents: string[], message: string): Promise<api.IVersion> {
+    public async write(tree: api.ITree, parents: string[], message: string, ref: string): Promise<api.IVersion> {
         return Promise.reject(new Error("Invalid operation"));
     }
