+++ packages/test/snapshots/package.json
@@ -54,7 +54,14 @@
   "dependencies": {
     "@fluid-internal/replay-tool": "^0.35.0",
     "@fluidframework/common-utils": "^0.27.0",
-    "assert": "^2.0.0"
+    "@fluidframework/driver-definitions": "^0.34.0",
+    "@fluidframework/driver-utils": "^0.34.0",
+    "@fluidframework/file-driver": "^0.34.0",
+    "@fluidframework/merge-tree": "^0.34.0",
+    "@fluidframework/protocol-definitions": "^0.1019.0-0",
+    "@fluidframework/replay-driver": "^0.34.0",
+    "assert": "^2.0.0",
+    "jsonschema": "^1.2.6"
   },
   "devDependencies": {
     "@fluidframework/build-common": "^0.20.0-0",
