+++ packages/test/snapshots/src/replayMultipleFiles.ts
@@ -9,6 +9,7 @@ import nodePath from "path";
 import { ReplayArgs, ReplayTool } from "@fluid-internal/replay-tool";
 import { Deferred } from "@fluidframework/common-utils";
 import { pkgVersion } from "./packageVersion";
+import { validateSnapshots } from "./validateSnapshots";
 
 // Determine relative file locations
 function getFileLocations(): [string, string] {
@@ -36,6 +37,7 @@ export enum Mode {
     Compare, // Compare to files stored on disk
     Stress,  // Do stress testing without writing or comparing out files.
     Validate,
+    BackCompat,
 }
 
 export interface IWorkerArgs {
@@ -77,7 +79,15 @@ class ConcurrencyLimiter {
     }
 }
 
-export async function processOneFile(args: IWorkerArgs) {
+/**
+ * Processes one node of the snapshot contents folder. Each node has the following:
+ * 1. messages.json - This file contains the messages to be replayed by the replay tool.
+ * 2. A list of snapshots in the current version. When running in "Compare" mode, snapshots generated by the replay
+ *    tool are compared against these.
+ * 3. original_snapshots - This optional folder contains a list snapshots in a previous version. When running in
+ *    "Validate" mode, a container is loaded from each of these snapshot and ensured that it loads successfully.
+ */
+export async function processOneNode(args: IWorkerArgs) {
     const replayArgs = new ReplayArgs();
 
     replayArgs.verbose = false;
@@ -153,6 +163,9 @@ export async function processContent(mode: Mode, concurrently = true) {
             case Mode.Write:
                 await processNodeForWrite(data, concurrently, limiter);
                 break;
+            case Mode.BackCompat:
+                await processNodeForBackCompat(data);
+                break;
             default:
                 await processNode(data, concurrently, limiter);
         }
@@ -171,19 +184,19 @@ async function processNodeForValidate(
     limiter: ConcurrencyLimiter,
 ) {
     // The snapshots in older format are in "src_snapshots" folder.
-    const srcSnapshotsFolder = `${data.folder}/src_snapshots`;
-    if (!fs.existsSync(srcSnapshotsFolder)) {
+    const srcSnapshotsDir = `${data.folder}/src_snapshots`;
+    if (!fs.existsSync(srcSnapshotsDir)) {
         return;
     }
 
-    // Each sub-directory under "src_snapshots" folder contain snapshots from a particular version. Process each one
+    // Each sub-directory under "src_snapshots" folder contain snapshots from an older version. Process each one
     // of these folders.
-    for (const node of fs.readdirSync(srcSnapshotsFolder, { withFileTypes: true })) {
+    for (const node of fs.readdirSync(srcSnapshotsDir, { withFileTypes: true })) {
         if (!node.isDirectory()) {
             continue;
         }
 
-        data.initializeFromSnapshotsDir = `${srcSnapshotsFolder}/${node.name}`;
+        data.initializeFromSnapshotsDir = `${srcSnapshotsDir}/${node.name}`;
         await processNode(data, concurrently, limiter);
     }
 }
@@ -229,6 +242,39 @@ async function processNodeForWrite(
 }
 
 /**
+ * Runs back-compat snapshot tests. We have a set of reference snapshots. Each set has snapshots in the current format
+ * and may have snapshots in an older format. For each set, the test does the following:
+ * 1. Loads a document with the snapshot in old version.
+ * 2. Takes a snapshot of the document.
+ * 3. Validates that the snapshot matches with the corresponding snapshot in current version.
+ * 4. Loads a document with snapshot in current version. Repeats steps 2 and 3.
+ */
+async function processNodeForBackCompat(data: IWorkerArgs) {
+    // Snapshots in current format is under "current_snapshots" directory.
+    const currentSnapshotsDir = `${data.folder}/current_snapshots`;
+
+    // Validate that we can load snapshot in current format and write it in current snapshot format.
+    await validateSnapshots(currentSnapshotsDir, currentSnapshotsDir);
+
+    // Snapshots in old format are under "src_snapshots" directory. If we don't have any, there is nothing more to do.
+    const srcSnapshotsDir = `${data.folder}/src_snapshots`;
+    if (!fs.existsSync(srcSnapshotsDir)) {
+        return;
+    }
+
+    // Each sub-directory under "src_snapshots" folder contain snapshots from an older version. Process each one
+    // of these folders.
+    for (const node of fs.readdirSync(srcSnapshotsDir, { withFileTypes: true })) {
+        if (!node.isDirectory()) {
+            continue;
+        }
+
+        // Validate that we can load snapshot from this older version format and write it in current snapshot format.
+        await validateSnapshots(`${srcSnapshotsDir}/${node.name}`, currentSnapshotsDir);
+    }
+}
+
+/**
  * Process one folder from the reference snapshots. It creates a worker thread and assigns the processing work to
  * the threads. If concurrently if false, directly processes the snapshots.
  */
@@ -246,7 +292,7 @@ async function processNode(
     }
 
     if (!concurrently || !threads) {
-        await processOneFile(data);
+        await processOneNode(data);
         return;
     }
