+++ packages/test/snapshots/src/replayWorker.ts
@@ -4,10 +4,10 @@
  */
 
 import threads from "worker_threads";
-import { IWorkerArgs, processOneFile } from "./replayMultipleFiles";
+import { IWorkerArgs, processOneNode } from "./replayMultipleFiles";
 
 const data = threads.workerData as IWorkerArgs;
-processOneFile(data)
+processOneNode(data)
     .catch((error) => {
         // eslint-disable-next-line no-null/no-null
         if (typeof error === "object" && error !== null && (error as Error).message !== undefined) {
