+++ packages/tools/replay-tool/src/index.ts
@@ -4,4 +4,5 @@
  */
 
 export * from "./replayArgs";
+export * from "./helpers";
 export * from "./replayMessages";
